+++ lib/excel2Json.js
@@ -8,7 +8,7 @@ function ExcelToJSON(logger) {
 }
 
 ExcelToJSON.prototype.readExcelFile = function(file, cb) {
- console.log(' ******** Reading EXCEL file ******** ' );
+ this.logger.info(' ******** Reading EXCEL file ******** ' );
  return XLSX.readFile(file);
 }
